Refactor and restructure end_model/ to classification/
