+++ snorkel/classification/training/__init__.py
@@ -1,0 +1,7 @@
+from .loggers import (  # noqa: F401
+    Checkpointer,
+    LogManager,
+    LogWriter,
+    TensorBoardWriter,
+)
+from .trainer import Trainer  # noqa: F401
