+++ snorkel/classification/training/schedulers/shuffled_scheduler.py
@@ -1,7 +1,7 @@
 import random
 from typing import Any, Dict, Sequence, Tuple
 
-from snorkel.end_model.data import MultitaskDataLoader
+from snorkel.classification.data import MultitaskDataLoader
 
 from .scheduler import Scheduler
